4/11/23 - 3:08PM-EST
-Added TMP (TextMeshPro)
-Added a working health bar
-Added a working timer
-Added Totem Icons

+++ Platformer_test/Assets/Scripts/Enemy/EnemyDamage.cs
@@ -6,7 +6,7 @@ public class EnemyDamage : MonoBehaviour
 {
     public PlayerHealth playerHealth;
 
-    public int damage = 10;
+    public float damage = 10.0f;
     // Start is called before the first frame update
     void Start()
     {
